Added if condition to check if the file exists

+++ logman.sh
@@ -7,9 +7,17 @@ cd /var/log/
 echo "This is the content of the original logs"
 echo ""
 
+if [ -f /var/log/vmware-vmsvc-root.log ]; then
 
-cat /var/log/vmware-vmsvc-root.log
-cp /var/log/vmware-vmsvc-root.log /var/log/vmware-root.log.old
+	# File exists
 
-truncate -s 0 vmware-vmsvc-root.log
+	cat /var/log/vmware-vmsvc-root.log
+	cp /var/log/vmware-vmsvc-root.log /var/log/vmware-root.log.old
 
+	truncate -s 0 vmware-vmsvc-root.log
+else
+	# File does not exist
+	echo 'The log file does not exist'
+
+
+fi
